backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
llama-cpp-python==0.2.24
transformers==4.38.2
torch==2.1.2
sentencepiece==0.1.99
tqdm==4.66.1
antlr4-python3-runtime==4.11.1
pygraphviz==1.10
networkx==3.1
jinja2==3.1.2
